traffic_analytics_app/requirements.txt
Cython
matplotlib>=3.2.2
numpy>=1.18.5
Pillow
PyYAML>=5.3
scipy>=1.4.1
opencv-python>=4.1.2
Pillow
tqdm>=4.41.0
requests==2.26.0
pyyaml==5.4.1
easydict
seaborn>=0.11.0
pandas
ftfy==6.0.3
psycopg2-binary
